[build] tt-forge-fe wheel
This change enables creating a `tt-forge-fe` wheel. To create a wheel,
run either of the following commands (inside our `venv`):
- `python setup.py bdist_wheel`
- `python -m build wheel` - this one requires build to be installed,
  `pip install build`

Logic for creating the wheel is contained in `setup.py`. It simply runs
cmake to build and install the project. The `setuptools` then picks up
and creates the wheel.

`tt-mlir` is now an `ExternalProject` in cmake.

Installing `tt-mlir` via `cmake --install build --component SharedLib`
installs all needed dependencies for `tt-mlir` (including `tt-metal`).

`TT_METAL_HOME` environment variable is no longer needed, small piece of
logic in `forge/forge/__init__.py` tries to deduce it, relative to its
location. If it can't be deduced, an error message will be shown.

However, if the `TT_METAL_HOME` variable is set the logic will respect
it as long as it is a valid path, so overriding it in custom scenarios
is possible.

Changes to the CI workflows will follow this change.

Issue #200