Update build.md with readability and grammar changes
Prerequisites - should not be numbered unless you are required to do those steps in order.
Some changes were to add "articles" (a, an, the). It's common with English as a Second Language authors to miss those because it is not native to all languages (for example, Russian).
A couple of sentences had the action at the end of the sentence. You should always lead with the action, then give the reason for the action.
I am Ben Moore and I am contributing to Tenstorrent documentation to enhance my portfolio. If you like what I did here, I am currently interviewing for multiple Technical Writer contracts at Tenstorrent and would appreciate you letting hiring managers know what you liked about my suggestions. 
https://bikerben33.github.io/portfolio/